rename nnet itf
