+++ speechx/speechx/nnet/nnet_itf.h
